Migrate skill configuration POST endpoint to /api/w/[wId]/skills
