+++ front/components/skill_builder/submitSkillBuilderForm.ts
@@ -1,7 +1,7 @@
 import type { SkillBuilderFormData } from "@app/components/skill_builder/SkillBuilderFormContext";
 import { clientFetch } from "@app/lib/egress/client";
-import type { PostSkillConfigurationResponseBody } from "@app/pages/api/w/[wId]/assistant/skill_configurations";
 import type { PatchSkillConfigurationResponseBody } from "@app/pages/api/w/[wId]/assistant/skill_configurations/[sId]";
+import type { PostSkillConfigurationResponseBody } from "@app/pages/api/w/[wId]/skills";
 import type { Result, UserType, WorkspaceType } from "@app/types";
 import { Err, normalizeError, Ok } from "@app/types";
 
@@ -25,7 +25,7 @@ export async function submitSkillBuilderForm({
   try {
     const endpoint = skillConfigurationId
       ? `/api/w/${owner.sId}/assistant/skill_configurations/${skillConfigurationId}`
-      : `/api/w/${owner.sId}/assistant/skill_configurations`;
+      : `/api/w/${owner.sId}/skills`;
 
     const method = skillConfigurationId ? "PATCH" : "POST";
