+++ front/pages/api/w/[wId]/skills/index.test.ts
@@ -2,11 +2,19 @@ import type { RequestMethod } from "node-mocks-http";
 import { describe, expect, it } from "vitest";
 
 import { Authenticator } from "@app/lib/auth";
+import {
+  SkillConfigurationModel,
+  SkillMCPServerConfigurationModel,
+} from "@app/lib/models/skill";
+import { MCPServerViewResource } from "@app/lib/resources/mcp_server_view_resource";
 import { SkillConfigurationResource } from "@app/lib/resources/skill/skill_configuration_resource";
 import { AgentConfigurationFactory } from "@app/tests/utils/AgentConfigurationFactory";
 import { FeatureFlagFactory } from "@app/tests/utils/FeatureFlagFactory";
 import { createPrivateApiMockRequest } from "@app/tests/utils/generic_private_api_tests";
+import { MCPServerViewFactory } from "@app/tests/utils/MCPServerViewFactory";
+import { RemoteMCPServerFactory } from "@app/tests/utils/RemoteMCPServerFactory";
 import { SkillConfigurationFactory } from "@app/tests/utils/SkillConfigurationFactory";
+import { SpaceFactory } from "@app/tests/utils/SpaceFactory";
 import type { MembershipRoleType } from "@app/types";
 import type {
   SkillConfigurationRelations,
@@ -361,9 +369,135 @@ describe("GET /api/w/[wId]/skills?withRelations=true", () => {
   });
 });
 
+describe("POST /api/w/[wId]/skills", () => {
+  it("creates a simple skill configuration", async () => {
+    const { req, res, workspace } = await setupTest("POST", "admin");
+
+    req.body = {
+      name: "Simple Skill",
+      description: "A simple skill without tools",
+      instructions: "Simple instructions",
+      tools: [],
+    };
+
+    await handler(req, res);
+    expect(res._getStatusCode()).toBe(200);
+
+    const responseData = res._getJSONData();
+    expect(responseData.skillConfiguration).toMatchObject({
+      name: "Simple Skill",
+      description: "A simple skill without tools",
+      instructions: "Simple instructions",
+      status: "active",
+      version: 0,
+      tools: [],
+    });
+
+    // Verify skill was created in the database
+    const skillConfiguration = await SkillConfigurationModel.findOne({
+      where: {
+        workspaceId: workspace.id,
+        name: "Simple Skill",
+      },
+    });
+    expect(skillConfiguration).not.toBeNull();
+  });
+
+  it("creates a skill configuration with 2 tools", async () => {
+    const { req, res, workspace, authenticator, user } = await setupTest(
+      "POST",
+      "admin"
+    );
+
+    // Create spaces (system space is required for MCP servers)
+    await SpaceFactory.system(workspace);
+
+    const globalSpace = await SpaceFactory.global(workspace);
+
+    const server1 = await RemoteMCPServerFactory.create(workspace, {
+      name: "Server 1",
+    });
+    const server2 = await RemoteMCPServerFactory.create(workspace, {
+      name: "Server 2",
+    });
+
+    const serverView1 = await MCPServerViewFactory.create(
+      workspace,
+      server1.sId,
+      globalSpace
+    );
+    const serverView2 = await MCPServerViewFactory.create(
+      workspace,
+      server2.sId,
+      globalSpace
+    );
+
+    req.body = {
+      name: "Test Skill",
+      description: "A test skill description",
+      instructions: "Test instructions for the skill",
+      tools: [
+        { mcpServerViewId: serverView1.sId },
+        { mcpServerViewId: serverView2.sId },
+      ],
+    };
+
+    await handler(req, res);
+    expect(res._getStatusCode()).toBe(200);
+
+    const responseData = res._getJSONData();
+    expect(responseData.skillConfiguration).toMatchObject({
+      name: "Test Skill",
+      description: "A test skill description",
+      instructions: "Test instructions for the skill",
+      status: "active",
+      version: 0,
+      tools: [
+        { mcpServerViewId: serverView1.sId },
+        { mcpServerViewId: serverView2.sId },
+      ],
+    });
+
+    // Verify skill was created in the database
+    const skillConfiguration = await SkillConfigurationModel.findOne({
+      where: {
+        workspaceId: workspace.id,
+        name: "Test Skill",
+      },
+    });
+    expect(skillConfiguration).not.toBeNull();
+    expect(skillConfiguration!.description).toBe("A test skill description");
+    expect(skillConfiguration!.instructions).toBe(
+      "Test instructions for the skill"
+    );
+    expect(skillConfiguration!.authorId).toBe(user.id);
+
+    // Verify tools were created in the database
+    const toolConfigurations = await SkillMCPServerConfigurationModel.findAll({
+      where: {
+        workspaceId: workspace.id,
+        skillConfigurationId: skillConfiguration!.id,
+      },
+    });
+    expect(toolConfigurations).toHaveLength(2);
+
+    const serverViewIds = toolConfigurations.map((t) => t.mcpServerViewId);
+    const view1 = await MCPServerViewResource.fetchById(
+      authenticator,
+      serverView1.sId
+    );
+    const view2 = await MCPServerViewResource.fetchById(
+      authenticator,
+      serverView2.sId
+    );
+    expect(serverViewIds).toContain(view1!.id);
+    expect(serverViewIds).toContain(view2!.id);
+  });
+});
+
 describe("Method Support /api/w/[wId]/skills", () => {
   it("should return 405 for unsupported methods", async () => {
-    for (const method of ["POST", "PUT", "PATCH", "DELETE"] as const) {
+    for (const method of ["PUT", "PATCH", "DELETE"] as const) {
       const { req, res, workspace } = await setupTest(method);
 
       req.query = { ...req.query, wId: workspace.sId };
@@ -374,7 +508,7 @@ describe("Method Support /api/w/[wId]/skills", () => {
       expect(res._getJSONData()).toMatchObject({
         error: {
           type: "method_not_supported_error",
-          message: "The method passed is not supported, GET is expected.",
+          message: "The method passed is not supported, GET or POST expected.",
         },
       });
     }
